Implemented computation of the matrix g for the exchange, so that the exchange field can be computed via a matrix (g) vector (m) multiplication, i.e. Hex = g*m rather than Hex = assemble(dE_dM).
I suggest to call the method we used so far 'box-assemble' and the new method 'box-matrix', as
both use the box method, but different ways of computing H_ex.

The current 'box-matrix' implementation uses a numpy array matrix to store g. This is not ideal
as it is stored as a full matrix where g is actually sparse. By moving the matrix
to a linear algebra backend that supports sparse matrices (such as the dolfin-supported
PETSC), this can be addressed.

The current 'box-matrix' implementation results in a speed up of all regression tests in
finmag/src from 158 seconds to 96 seconds (on my virtual ubuntu machine with 1GB RAM, running
on a MacBook Pro). I have thus made this the default method.

I will commit another file in a moment that can be used to test the performance ratio of
box-matrix over max-assemble.

This commit includes a test (in test_exchange) which checks that the box-assemble and the
box-matrix method give identical results (apart from numerics).


[#25041739]