Testing idea of box-matrix method computation, and measuring performance relative to box-assemble.
File that (i) tests the basic idea of the box-matrix method (probably superseeded by test
in src/finmag/sim/tests/test_exchange.py) and (ii) shows how the speed up over the box-assemble
method decreases as the mesh size increases:

Speed up (larger than 1 means matrix method is faster than assembly)
  1 -> 69.35
  5 -> 82.77
 10 ->  8.54
 15 ->  1.11
 20 ->  0.45

The number n on the left is the number of simplices in one direction for a 2d mesh.
The number of rows in the matrix is thus n^2 (because of the 2d mesh), and the number
of matrix entries is (n^2)^2.

The performance should get better when we move to sparse storage of that matrix.