Add running of relevant test to get image files ready for documentation.
Not sure this is ideal: in the current setup we need to maintain the list of
required files in this Makefile, but will not run code unneccesarily, and
'make html', for example, will 'just work'.

Alternatively, we could always run py.test on ../examples, but this will
take longer, and maybe forgotten. Let's see how it goes.