Tidying up the llg class a little (set_m0)
Calling set_m0 a second time, will not override the self._m object
but update it's values.

This also defines _m the moment when we create a class, which is good
as it then doesn't matter in which order we call further methods of the
llg class: the central _m object is present from the beginning.

Removed associated test that documented undesirable behaviour.