Modifying Makefile to create files in examples/time-dependent-applied-field/ that are included in documentation.
In theory, py.test should do this, and it seems to do it fine for the macrospin
example. However, the output file for the time-dependent-applied field example
are stored not in the directory of this example, but at the top level where
py.test is called.

This seems odd, and possibly py.test can be told how to do this differently
(but I didn't find anything right now).

For now, we go the explicit way of creating the files when we compile
the documentation, if the files are missing.