Moving Weiwei's comparison with Magpar (exchange and anisotropy) into src/finmag/sim/tests so that they are executed automatically.
This means that a working magpar needs to be installed on all sytems to
pass all unit tests. THis is easily done with the script in
install/magpar.sh

The tolerances set here are of the order of 1-e5. This is a limitation
of the inp file format that is used to store the data (as only 6 digits
or so are stored in this text-based format).