Modified version of Anders draft for convergence test:
- ignores Nmag for now (but should get this back in again)

- runs much faster

Some observations:

- we currently load the netgen mesh. Stricly, we should scale the
coordinates by 10e-9 (see commented line), but this makes the H_demag
field worse. Presumably, one of the solvers uses an absolute tolerance
that affects the calculation here. This means two things:

(i) it may be necessary to rescale the coordinates for solving for phi to
something of order of unity

(ii) we need to explore the settings for the numerical solvers to see
whether the results are affected by this.