Made REL_TOLERANCE values stricter, and used different ones for every test.
I don't understand why the 3d example shows relatively poor accuracy.

I also observer that -- for exampel in the 1d example -- if we increase
the OOMMF mesh resolution, the relative error increases. This seems odd
as well. On the positive side, by increasing the number of FE nodes to 100,000,
we get a relative error of less that 2e-9.

As the agreement with Magpar and Weiwei's FE code is good for the anisotropy
this is more likely to be a matter of either the code we use to drive OOMMF
or the particular configurations we test. However, that is just guess work
for now.