Moved the DMI class to be now derived from the new EnergyBase class thus sharing a lot of code with Zeeman and Exchange class -- this was replication of a lot of complicated code before.
As we don't have that many tests for the skyrmions, I have preserved
the original class under the name DMI_Old, and extended some tests
to test the new and the old class.

The vision is to get rid of the old class soon.

I also noted that most likely there is a factor mu0 * Ms missing in
the definition of the DMI energy. We need to divide by mu0*Ms when
we compute the field (which happens automatically in EnergyBase now).
I have added the mu0*Ms to the definition of the DMI-energy (but not
in DMI_Old). It is of course possible that this factor is already
absorbed in the parameter D. In this case, we should redefine the use
of D in our simulations, so that the current D changes to D/(mu0*Ms).

As long as our material is permalloy like, this does not matter that
much as Ms*mu0 is approximately 1.