Extracted functions from calc_llg_dmdt.
Decided how to do so with the help of the committed benchmarks.
Looping over the nodes several times is prohibitive, now simply
extracted the computation of a term for one node at a time.
This doesn't cost performance and reduces copy & paste between
different variants of our LLG(B,S,...) equation.

Roadmap:
- Extract Slonczewski into own function as well.
- Do the same for the Jacobian.
- Can the code be simplified/unified any further?