Some suggestions for consideraton: - DG0 allows discontinuous steps; maybe better than CG1 space. - interpolate is better than project as it interpolates the expression at the nodes (project   will minimse the L2 norm of the function deviation and may thus result in unexpected behaviour here)
Final note of caution: the success of these changes is difficult to see as any VTK-based
visualisation uses linear interpolation on the mesh, despite the DG0 space being constant in every
element. To see this, one can create a second (much) finer mesh and project or interpolate the course DG0 funcction on the fine mesh to see the constantness in every element (run out of time here).