Adding some notes on parallel execution. Dmitri and all, if you see anything wrong here, please correct -- I am working from memory.
This has been triggered (quite rightly) by Max' comment and observation on pivotal tracker:

[#35772117]

Hello,

Maximilian Albert has added a feature 'Finmag and parallel execution' to the project 'FinMag'.

It would be nice to be run Finmag simulations in parallel on multiple processors. I have no idea if this currently works (or has even been tried). However, I just stumbled upon two lines of code in sim/llg.py and llb/material.py, which showed a revealing difference:

 sim/llg.py:  self._m.vector()[:] = ...

 llb/material.py:  self._m.vector().set_local(...)

Again, I have no idea whether these two lines do the same thing (or even whether either of them would do the "right" thing) when executed in parallel. But it would be good to get a general overview of the "FEniCS idioms" that we need to stick to if we ever want to be able to run Finmag in parallel. So this is more of a reminder to start thinking about this at some point, not a request to make parallel execution work immediately.

http://www.pivotaltracker.com/story/show/35772117