fix "Illegal Instruction" problem with binary of finmag (with Dmitri)
The compilation optimised for the native CPU (corei7 on my machine).
Osiris and Max's machine run older CPUs (most likely core2), which don't support
some of the instruction sets on my machine. This change
has been tested to work on osiris. In the near future, we should
introduce an option to the Makefile for compilation for binary distribution.