Changed names a little to make it (at least for me) clearer what happens.
Looking at the results, it seems the dolfin expression for no
spatial dependence are much slower than the numpy expression. This could come through the use of the fill() method in 'numpy_vectorised_no_x_dependence()'
which exploits our extra knowledge about the lack of x-dependence and
thus is fast.

It would be nice to compare a vectorised numpy version that includes the x-dependence with the spatially varying x. Can this be done with numpy?