finished new timings report, added ipython notebook that demos it
report now shows not only the method/functions that took the longest time, but
also a breakdown of runtime by class/module. It detects time that was not
accounted for as 'untimed' or if multiple timings were taken at the same time
and recorded time is higher than wall time, reports it as 'redundant timings'.

ipython notebook demos this new table and shows a plot of the breakdown by
class/module