New tutorial notebook (still in semi-draft mode) with a couple of relaxations of a single nanodisk. Demonstrates a few issues with time integration we might want to think about:
Interestingly, in the second relaxation dmdt keeps increasing a lot of times,
but only by a tiny amount. This caused the simulation to finish prematurely.
With a higher default value of dmdt_increased_counter_limit it now reaches
relaxation, but in both simulations it turns out that relaxation is actually
reached quite soon and then M keeps creeping veeery slowly until dmdt is below
the tolerance. It would be good to discuss if we can find a better solution
for these issues.