Add create_integrator method to simulation class
This is useful if we want to set integrator parameters: while the
integrator object has methods to set tolerances, the integrator object
is only created once advance_time is called. This function solves the
matter. It also allows to pass parameters like abs_tol and rel_tol
directly to the integrator when it is constructed.

With Max and Marijan.