Field will now interpolate from one FunctionSpace to another
when setting Field values with another Field with a different FunctionSpace,
the Field class would just overwrite the original function. That means the
original functionspace and the new function would not match! This has now
been rectified by letting the Field class interpolate the values given. The
decision to interpolate and not project is backed by unit tests.