Remove nmag test
The error originates from the new mesh have 1 node less than the
original mesh used for testing:

test_cubic_anis_nmag.py:13: in test_cubic_against_nmag
    assert m_ref.shape == m_computed.shape
E   assert (1538, 3) == (1537, 3)
E     At index 0 diff: 1538 != 1537
E     Full diff:
E     - (1538, 3)
E     ?     ^
E     + (1537, 3)
E     ?     ^

We speculate that this is from a new version of netgen, relative
to the tests.

The Nmag test code is not available, but the results stored as a text
file, so we cannot easily update the results. As we have a large number
of other tests (and a working comparison with magpar), we remove this
test now.