remove nmag comparison test
The error originates from the new mesh being slightly different
from the old mesh for which the test reference data was computed.

We speculate that this is from a new version of netgen, relative
to the tests.

The Nmag test code is not available, but the results stored as a text
file, so we cannot easily update the results. As we have a large number
of other tests (and a working comparison with magpar), we remove this
test now.